+++ Makefile
@@ -48,6 +48,7 @@ INCS	+=	$(INCD)/keys.h
 SRCD	=	srcs
 
 SRCS	=	$(SRCD)/main.c
+SRCS	+=	$(SRCD)/get_map.c
 SRCS	+=	$(SRCD)/check_map.c
 SRCS	+=	$(SRCD)/window.c
