+++ srcs/check_map.c
@@ -5,14 +5,14 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: gudias <marvin@42lausanne.ch>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/08/18 13:42:33 by gudias            #+#    #+#             */
-/*   Updated: 2022/08/18 17:20:28 by gudias           ###   ########.fr       */
+/*   Created: 2022/08/18 17:40:48 by gudias            #+#    #+#             */
+/*   Updated: 2022/08/18 17:48:02 by gudias           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-static int	check_extension(char *mapname)
+int	check_extension(char *mapname)
 {
 	int	len;
 
@@ -20,49 +20,4 @@ static int	check_extension(char *mapname)
 	if (ft_strncmp(mapname + (len - 4), ".cub", 4) != 0)
 		return (1);
 	return (0);
-}
-
-static int	open_map(char *mapname)
-{
-	int	fd;
-	char	*path;
-
-	path = ft_strjoin(MAPSDIR, mapname);
-	fd = open(path, O_RDONLY);
-	free(path);
-	return (fd);
-}
-
-static char	**read_map(int fd)
-{
-	char	*line;
-
-	line = get_next_line(fd);
-	while (line)
-	{
-		ft_printf("%s", line);
-		free(line);
-		line = get_next_line(fd);
-	}
-	return ((char**)1);
-}
-
-char	**get_map(char *mapname)
-{
-	int	fd;
-	char	**map;
-
-	if (check_extension(mapname))
-	{
-		ft_printf("invalid map extension, must be .cub\n");
-		return (NULL);
-	}
-	fd = open_map(mapname);
-	if (fd <= 0)
-	{
-		ft_printf("couldn't find map in %s\n", MAPSDIR);
-		return (NULL);
-	}
-	map = read_map(fd);
-	return (map);
 }
